Small changes and added README

+++ src/main/java/codenameburkley/DiskData.java
@@ -3,9 +3,9 @@ package codenameburkley;
 public class DiskData {
     public String model;
     public String serial;
-    public long sizeBytes;
-    public long readBytes;
-    public long writeBytes;
+    public double sizeBytes;
+    public double readBytes;
+    public double writeBytes;
     public long readCount;
     public long writeCount;
     public long transferTimeMs;
